gate retry test data to xcode 13

+++ prepareTestResults.sh
@@ -13,23 +13,25 @@ xcodebuild test \
     -skip-testing:SampleAppUITests/RetryTests \
     -resultBundlePath "$FILENAME" || true
 
-# "Mixed" test results must be run separately to use -retry-tests-on-failure
-RETRY_FILENAME='RetryResults.xcresult'
-rm -rf "$RETRY_FILENAME"
-xcodebuild test \
-    -project SampleApp.xcodeproj \
-    -scheme MainScheme \
-    -destination 'platform=iOS Simulator,name=iPhone 8,OS=latest' \
-    -test-iterations 2 \
-    -retry-tests-on-failure \
-    -only-testing:SampleAppUITests/RetryTests \
-    -resultBundlePath "$RETRY_FILENAME" || true
-
 echo "Even if some test failed this is OK."
 
 echo "${FILENAME} should contain succeed, failed and skipped tests for xchtmlreport functional testing"
 rm -rf "../Tests/XCTestHTMLReportTests/${FILENAME}"
 mv "$FILENAME" "../Tests/XCTestHTMLReportTests/"
+
+if [[ $XCODE_VERSION != 12.* && $XCODE_VERSION != 11.* ]]; then
+    # "Mixed" test results must be run separately to use -retry-tests-on-failure
+    RETRY_FILENAME='RetryResults.xcresult'
+    rm -rf "$RETRY_FILENAME"
+    xcodebuild test \
+        -project SampleApp.xcodeproj \
+        -scheme MainScheme \
+        -destination 'platform=iOS Simulator,name=iPhone 8,OS=latest' \
+        -test-iterations 2 \
+        -retry-tests-on-failure \
+        -only-testing:SampleAppUITests/RetryTests \
+        -resultBundlePath "$RETRY_FILENAME" || true
+fi
 
 echo "${RETRY_FILENAME} will contain mixed test results"
 rm -rf "../Tests/XCTestHTMLReportTests/${RETRY_FILENAME}"
